Make theano (for pymc3) optional dependency (#356)
* fix docs

* fix2

* make theano rtd req

* catch optional deps

* remove inherited members, restrict specials, from docs

* Delete hierarchical.ipynb

* simpliy autodoc

* inc sphinx v

* test

* test

* test

* test

* test

* restructure

* fixup

* fixup2
