+++ doc/api_engine.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.engine
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
