+++ doc/api_logging.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.logging
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
