+++ doc/api_objective.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.objective
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
