+++ doc/api_optimize.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.optimize
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
