+++ doc/api_petab.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.petab
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
