+++ doc/api_problem.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.problem
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
