+++ doc/api_profile.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.profile
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
