+++ doc/api_result.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.result
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
