+++ doc/api_sampling.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.sampling
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
