+++ doc/api_startpoint.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.startpoint
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
