+++ doc/api_storage.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.storage
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
