+++ doc/api_visualize.rst
@@ -1,5 +1,1 @@
 .. automodule:: pypesto.visualize
-   :members:
-   :inherited-members:
-   :special-members:
-   :imported-members:
