+++ doc/conf.py
@@ -25,8 +25,7 @@ sys.path.insert(0, os.path.abspath('../'))
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
-#
-# needs_sphinx = '1.0'
+needs_sphinx = '3.0.4'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -39,10 +38,13 @@ extensions = ['sphinx.ext.autodoc',
 
 # default autodoc options
 # list for special-members seems not to be possible before 1.8
-autodoc_default_flags = ['members',
-                         'undoc-members',
-                         'show-inheritance',
-                         'imported-members']
+autodoc_default_options = {
+    'members': True,
+    'undoc-members': True,
+    'special-members': '__init__, __call__',
+    'imported-members': True,
+    'show-inheritance': True,
+}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
