+++ readthedocs.yml
@@ -1,4 +1,4 @@
 build:
     image: latest
 python:
-    version: 3.6
+    version: 3.8
